fix: Resolve all remaining MEDIUM+LOW bugs — 115 fixes across 30 files
8 parallel agents swept all MEDIUM and LOW severity bugs from Walkthrough #5.

By category (83 MEDIUM, 32 LOW):

Memory safety (25M):
- Integer overflow guards in nimcp_malloc, aligned_alloc, backprop alloc
- realloc guard-byte corruption: compute real_ptr before system realloc
- VLA→heap in HYBRID mode backprop (missed in prior walkthrough)
- COW region table full: fail instead of returning unregistered region
- COW signal handler bounds check on source_page_idx
- QAT layer_configs leak on mutex creation failure
- GW competitor content leak on decay-prune (3 resolution paths)
- Exception double-free race: split new_count==0 vs <0

Correctness (22M):
- Weight statistics divisor: count valid handles, not total slots
- Timestamp underflow guards (calcium dynamics, synaptic traces)
- Immune bridge time source: real monotonic time, not simulated
- Query classifier word boundary checks (strncmp prefix-only bug)
- Single-word queries never hashed in word-boundary cue function
- Goto-past-initializer UB in concurrent reasoning finalize
- NLP uint16 truncation: bounds check before htons cast
- Topographic maps OOB write when cortical_dims==1
- Social memory spurious THROW on normal not-found lookup
- Batch loss accumulation skips error returns (-1.0f)

Thread safety (12M):
- Data races: GW get_ignition_threshold, SNC get_signal_type,
  SNC broadcast_dopamine, reasoning Portia-skip stats
- KG dispatch held modules_mutex during blocking enqueue
- Parallel stage pool_submit failures: mark tasks done to prevent hang

Python bindings (29M, 5L):
- 29 missing NULL checks on PyFloat/PyLong/PyUnicode return values
- Magic numbers replaced with named constants

Error handling (8M, 6L):
- Wrong error codes (NULL_POINTER→NO_MEMORY/IO) in 6 API functions
- Missing immune notification on snapshot save failure
- predict() resource loss on OOM: copy before freeing old decision
- Missing error handling on reasoning step additions
- fcntl/errno preservation in NLP socket setup

Code quality (21L):
- Dead code removal (unused computation, unreachable branches)
- Magic numbers→named constants (7 locations)
- Duplicate includes removed
- Consistent return codes (NIMCP_SUCCESS→NIMCP_OK)
- Type consistency (int→uint32_t loop vars, size_t for hash)
- Division-by-zero guards for tunable constants

All 32 convergent reasoning tests pass (23 unit + 7 integration + 2 e2e)
Score progression: 5.8 → 7.0 → 7.5 → 6.7 → 7.3 → ~8.0/10

+++ src/utils/containers/nimcp_hash_table.c
@@ -126,14 +126,14 @@ static uint32_t hash_murmur3_32(uint32_t h)
 static uint32_t hash_murmur3(const void* key, size_t key_size)
 {
     const uint8_t* data = (const uint8_t*) key;
-    const int nblocks = key_size / 4;
+    const size_t nblocks = key_size / 4;
     uint32_t h = 0;  // seed = 0
 
     const uint32_t c1 = 0xcc9e2d51;
     const uint32_t c2 = 0x1b873593;
 
     // Body - process 4-byte blocks (use memcpy to avoid unaligned access UB)
-    for (int i = 0; i < nblocks; i++) {
+    for (size_t i = 0; i < nblocks; i++) {
         uint32_t k;
         memcpy(&k, data + i * 4, sizeof(uint32_t));
 
@@ -163,8 +163,10 @@ static uint32_t hash_murmur3(const void* key, size_t key_size)
             h ^= k;
     }
 
-    // Finalization
-    h ^= key_size;
+    // Finalization — explicitly cast to uint32_t for MurmurHash3 spec compliance.
+    // On 64-bit systems, key_size is size_t (64-bit); the standard MurmurHash3
+    // finalization XORs with the 32-bit length value.
+    h ^= (uint32_t)key_size;
     return hash_murmur3_32(h);
 }
